requirements.txt
gunicorn==20.0.4
mysql-connector==2.2.9
mysql-connector-python==8.0.21
mysql-connector-python-rf==2.2.2
mysqlclient @ file:///C:/Users/blossom_shweta/Downloads/mysqlclient-1.4.6-cp38-cp38-win32.whl
Django==3.0.6
pytz==2020.1